requirements.txt
streamlit
PyPDF2
langchain
langchain_community
langchain_google_genai
faiss-cpu
google-generativeai
python-dotenv
arxiv
